+++ include/minishell.h
@@ -73,65 +73,75 @@ typedef struct s_env_var_ll
 	struct s_env_var_ll	*next;
 }				t_env_var_ll;
 
-//			FUNCTIONS
+//				FUNCTIONS
 
 void	minishell_error(const char *loc);
 
+//				FUNCTIONS: LEXER
 t_token		*lexer(const char *inp);
-void		get_token_info(const char *inp, size_t *pos, t_token *node);
 t_token_id	get_char_id(const char c);
 void		token_id_quote(const char *inp, size_t *pos, const t_token_id val);
 void		token_id_space(const char *inp, size_t *pos, const t_token_id val);
 void		token_id_misc(const char *inp, size_t *pos, const t_token_id val);
 
-void		mini_parse_input(char *input, t_cmd *cmd);
-char		**make_double_array(int word_count);
-int			ft_word_counter(char const *s, char c);
-void		free_double_array(char **double_array);
+//				FUNCTION: SYNTAX
 
-bool		check_builtin(char *cmd);
-void		execute_builtin(t_cmd *cmd, t_env_var_ll **our_env_var);
-int			execute_echo(t_cmd *cmd, int fd);
-int			execute_pwd(int fd);
-int			execute_cd(t_cmd *cmd, t_env_var_ll **env_var_list);
-void		print_env(t_env_var_ll *env_var_list);
-void		unset_env(char *name, t_env_var_ll **env_var_list);
-bool		env_var_exists(char *name, t_env_var_ll *env_var_list);
-char		*get_env(char *name, t_env_var_ll *env_var_list);
-void		set_env(char *envar, t_env_var_ll **env_var_list);
-void		execute_export(t_cmd *cmd, t_env_var_ll **env_var_list);
-void		print_export(t_env_var_ll *env_var_list);
-void		execute_exit(t_cmd *cmd);
+//				FUNCTIONS: PARSER
+void			get_token_info(const char *inp, size_t *pos, t_token *node);
+
+//				FUNCTIONS: EXPANDER
+
+//				FUNCTIONS: BUILTINS
+bool			builtin_check(char *cmd);
+void			builtin_execute(t_cmd *cmd, t_env_var_ll **our_env_var);
+int				builtin_echo(t_cmd *cmd, int fd);
+bool			builtin_echo_is_n_option(char *str);
+int				builtin_cd(t_cmd *cmd, t_env_var_ll **env_var_list);
+char			*builtin_cd_get_new_working_dir(t_cmd *cmd, t_env_var_ll **env_var_list);
+int				builtin_pwd(int fd);
+void			builtin_export(t_cmd *cmd, t_env_var_ll **env_var_list);
+void			builtin_export_print_export(t_env_var_ll *env_var_list);
+void			builtin_unset(char *name, t_env_var_ll **env_var_list);
+void			builtin_env(t_env_var_ll *env_var_list);
+void			builtin_exit(t_cmd *cmd);
+
+//				FUNCTIONS: INIT SHELL
+int				init_shell(char **envp, t_env_var_ll **env_var_list);
+void			init_shell_set_underscore(t_env_var_ll **env_var_list);
+void			init_shell_update_SHLVL(t_env_var_ll **env_var_list);
+int				init_shell_add_env_vars(char *env_var, t_env_var_ll **env_var_list);
+
+//				FUNCTIONS: HANDLE ENVIRONMENT VARIABLE LIST
+void			env_var_print_linked_list(t_env_var_ll *env_var_list); // FOR TESTING
+void			env_var_free_node(t_env_var_ll *env_var_list);
+t_env_var_ll	*env_var_init_new_var_node(char *env_var);
+int				env_var_add_to_end_list(t_env_var_ll **env_var_list, t_env_var_ll *new_var);
+
+//				FUNCTIONS: ENVIRONMENT VARIABLES
+bool			env_var_exists(char *name, t_env_var_ll *env_var_list);
+char			*env_var_get_env(char *name, t_env_var_ll *env_var_list);
+void			env_var_set_env(char *envar, t_env_var_ll **env_var_list);
+
+// 				FUNCTIONS: EXECUTOR
+
+// 				FUNCTIONS: TMP_MARES
+void			mini_parse_input(char *input, t_cmd *cmd);
+char			**make_double_array(int word_count);
+int				ft_word_counter(char const *s, char c);
+void			free_double_array(char **double_array);
 
 
-// OLD:
-// bool	get_env_var_exists(char *name); // old old, making use of getenv()
-// bool	env_var_exists(char *name, char **envp);
-// char	*get_env(char *name, t_env_var *envars); 
-// void	set_env(char *name, char *value, t_env_var *envars);
-// int		search_for_env_index(char *name, char **envp);
-// int		get_end_of_envp_list(char **envp);
-// char	*make_env_var_format(char *name, char *value);
-// void	set_our_envp(char **envp, t_env_var *envars);
+//					UTILS
 
-void			print_linked_list(t_env_var_ll *env_var_list); // FOR TESTING
-t_env_var_ll	*init_new_var(char *env_var);
-int				add_var_to_end_list(t_env_var_ll **env_var_list, t_env_var_ll *new_var);
-int				add_variable(char *env_var, t_env_var_ll **env_var_list);
-int				init_env_var(char **envp, t_env_var_ll **env_var_list);
+//					UTILS_TOKEN
 
+int			list_token_size(t_token *top);
+t_token		*list_token_new(void);
+t_token		*list_token_last(t_token *list);
+void		list_token_add_back(t_token **list, t_token *node);
 
-//				UTILS
-
-//				UTILS_TOKEN
-
-int		list_token_size(t_token *top);
-t_token	*list_token_new(void);
-t_token	*list_token_last(t_token *list);
-void	list_token_add_back(t_token **list, t_token *node);
-
-//		TEST FUNCTIONS (CAN BE REMOVED)
-void	list_token_print(t_token *top);
+//					TEST FUNCTIONS (CAN BE REMOVED)
+void		list_token_print(t_token *top);
 
 
 #endif
