+++ src/builtin/echo.c
@@ -8,7 +8,7 @@
 ** If -n is specified, the trailing newline is suppressed.
 */
 
-bool	is_n_option(char *str)
+bool	builtin_echo_is_n_option(char *str)
 {
 	int	i;
 	int	len;
@@ -29,7 +29,7 @@ bool	is_n_option(char *str)
 		return (false);
 }
 
-int	execute_echo(t_cmd *cmd, int fd)
+int		builtin_echo(t_cmd *cmd, int fd)
 {
 	int	i;
 	int	n_option;
