+++ src/builtin/old/builtin_old.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   builtin.c                                          :+:    :+:            */
+/*   builtin_old.c                                      :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: mverbrug <mverbrug@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/01/20 14:05:05 by mverbrug      #+#    #+#                 */
-/*   Updated: 2023/02/21 14:30:02 by mverbrug      ########   odam.nl         */
+/*   Updated: 2023/03/01 11:47:23 by mverbrug      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,3 +49,13 @@ void	execute_builtin(t_cmd *cmd, t_env_var *our_env_var)
 		// execute_exit();
 	
 }
+
+// ! include minishell.h
+// bool	get_env_var_exists(char *name); // old old, making use of getenv()
+// bool	env_var_exists(char *name, char **envp);
+// char	*get_env(char *name, t_env_var *envars); 
+// void	set_env(char *name, char *value, t_env_var *envars);
+// int		search_for_env_index(char *name, char **envp);
+// int		get_end_of_envp_list(char **envp);
+// char	*make_env_var_format(char *name, char *value);
+// void	set_our_envp(char **envp, t_env_var *envars);
